requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
Pillow==10.4.0
pdfplumber==0.11.4
PyMuPDF==1.24.10
pydantic==2.8.2
langchain==0.2.15
langchain-community==0.2.12
duckduckgo-search==6.2.11
faiss-cpu==1.8.0.post1
sentence-transformers==3.0.1
torch>=2.1.0
transformers==4.44.0
tools
fitz
frontend